feat: Add Docker Image CI workflow
Added a new GitHub Actions workflow file, `docker-image.yml`, to build a Docker image for the project. The workflow is triggered on pushes and pull requests to the `main` branch. It uses the latest version of Ubuntu and includes steps to checkout the code, build the Docker image using the `Dockerfile`, and tag it with a timestamp.

Deleted the `jekyll-gh-pages.yml` workflow file, as it is no longer needed for the project.